decomposed clusterizer to address compilation time issue (now reduced by factor x5)
