+++ RecoVertex/PrimaryVertexProducer_Alpaka/plugins/alpaka/ClusterizerAlgo.h
@@ -25,10 +25,18 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
 
   class ClusterizerAlgo {
   public:
-    ClusterizerAlgo(Queue& queue);
+    ClusterizerAlgo(Queue& queue, int32_t bSize);
+
     void clusterize(Queue& queue, portablevertex::TrackDeviceCollection& inputTracks, portablevertex::VertexDeviceCollection& deviceVertex, const std::shared_ptr<portablevertex::ClusterParamsHostCollection> cParams, int32_t nBlocks, int32_t blockSize); // Clusterization
+ 
+    void resplit_tracks(Queue& queue, portablevertex::TrackDeviceCollection& inputTracks, portablevertex::VertexDeviceCollection& deviceVertex, const std::shared_ptr<portablevertex::ClusterParamsHostCollection> cParams, int32_t nBlocks, int32_t blockSize); // Clusterization
+
+    void reject_outliers(Queue& queue, portablevertex::TrackDeviceCollection& inputTracks, portablevertex::VertexDeviceCollection& deviceVertex, const std::shared_ptr<portablevertex::ClusterParamsHostCollection> cParams, int32_t nBlocks, int32_t blockSize); // Clusterization    
     void arbitrate(Queue& queue, portablevertex::TrackDeviceCollection& inputTracks, portablevertex::VertexDeviceCollection& deviceVertex, const std::shared_ptr<portablevertex::ClusterParamsHostCollection> cParams, int32_t nBlocks, int32_t blockSize); // Arbitration
+  
   private:
+    cms::alpakatools::device_buffer<Device, double[]> beta_;
+    cms::alpakatools::device_buffer<Device, double[]> osumtkwt_;
   };
 
 }  // namespace ALPAKA_ACCELERATOR_NAMESPACE
